Reworked the pause menu
Redone the IMGUI pause menu.
It uses the usual canvas UI (Without TMP)

Preferences are also only saved if they changed.

Added Options to reduce the Post Processing Weight and to disable Auto-Exposure, which allows the player to get lighting closer to the 2012 release if they prefer it.

Moved the Option menu canvas into a prefab to simplify changes to scene.unity

As per request, the option menu is now inside a prefab and instantiated by "gui_skin_holder" on start.

Also removed the "settings" array and now searching for them through their parent instead. (Less setup and potential errors)