Reworked gun selection process
The gun selection option has a big negative side effect that basically locks people into never using the full arsenal of weapons available.

To counter this without removing it completely. The selection is only valid for the next round started.

To avoid confusion, the option has been moved into a new "restart menu" accessible via a button in the options.

Known "feature":
When selecting a weapon, the next round is started with it even without using the "restart" button.
(I like that behaviour and the fix wouldn't be all that pretty, so I'm calling it a feature)