Reworked trigger mechanisms
Instead of having various checks, we now check if it "is_connected".

This is closer to an actual firearm and allows for more room to work with.

It being connected implies that the firing sequence is stopped by the trigger.
When fully pulled, the trigger disconnects.
When firing semi automatic, the trigger instantly connects back.
When firing automatic, we only connect back when the trigger is released.